+++ src/value/mod.rs
@@ -1,17 +1,28 @@
 use std::fmt;
-use std::{cell::RefCell, rc::Rc};
-use std::str::FromStr;
+use std::rc::Rc;
 
-use super::evaluate;
-use super::parsetree::ParseTree;
 use super::errors::RuntimeError;
 
 pub mod string;
+pub mod exactlist;
+pub mod inductionlist;
+pub mod maplist;
+pub mod concatlist;
+pub use exactlist::ExactList;
+pub use inductionlist::LazyInductionList;
+pub use maplist::LazyMapList;
+pub use concatlist::LazyConcatList;
 
 #[derive(Clone)]
 pub enum Value {
     Number(i64),
     List(Rc<dyn ListLike>),
+}
+
+pub trait ListLike {
+    fn index(&self, i: i64) -> Result<Value, RuntimeError>;
+    fn length(&self) -> Result<i64, RuntimeError>;
+    fn force_resolve(&self) -> Result<(), RuntimeError>;
 }
 
 impl Value {
@@ -92,299 +103,9 @@ fn list_eq_helper(l1rc: &Rc<dyn ListLike>, l2rc: &Rc<dyn ListLike>) -> Result<bo
     Ok(false)
 }
 
-pub trait ListLike {
-    fn index(&self, i: i64) -> Result<Value, RuntimeError>;
-    fn length(&self) -> Result<i64, RuntimeError>;
-    fn force_resolve(&self) -> Result<(), RuntimeError>;
-}
-
-pub struct ExactList {
-    contents: Vec<Value>,
-}
-
-pub struct LazyInductionList {
-    function: ParseTree,
-    initial_value: Value,
-    resolved: RefCell<Vec<Value>>,
-}
-
-pub struct LazyMapList {
-    function: ParseTree,
-    source: Rc<dyn ListLike>,
-    resolved: RefCell<Vec<Option<Value>>>,
-}
-
-pub struct LazyConcatList {
-    first: Rc<dyn ListLike>,
-    second: Rc<dyn ListLike>,
-    firstlen: Option<i64>,
-}
-
-impl ExactList {
-    pub fn new(c: Vec<Value>) -> ExactList {
-        ExactList { contents: c }
-    }
-}
-
-impl ListLike for ExactList {
-    fn index(&self, i: i64) -> Result<Value, RuntimeError> {
-        let trueindex;
-        let len = self.length()?;
-
-        if i >= len || i < -len{
-            return Err(RuntimeError::OutOfBounds(format!(
-                "Attempted to access index {} of list of length {}",
-                i,
-                self.contents.len()
-            )));
-        }else if i >= 0{
-            trueindex = i;
-        }else{
-            trueindex = len+i;
-        }
-
-        let trueindex = usize::try_from(trueindex)
-            .map_err(|_| RuntimeError::OutOfBounds(format!("unknown error when indexing list (i = {})", i)))?;
-
-        Ok(self.contents[trueindex].clone())
-    }
-
-    fn length(&self) -> Result<i64, RuntimeError> {
-        return i64::try_from(self.contents.len())
-            .map_err(|_| RuntimeError::OutOfBounds(String::from("length could not be converted to i64")));
-    }
-
-    fn force_resolve(&self) -> Result<(), RuntimeError> {
-        for i in 0..self.contents.len(){
-            self.contents[i].force_resolve()?
-        }
-        Ok(())
-    }
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct ParseExactListError;
-
-impl FromStr for ExactList{
-    type Err = ParseExactListError;
-    fn from_str(s: &str) -> Result<Self, Self::Err>{
-        // Assert that the first and last chars are [...]
-        if s.chars().next().ok_or(ParseExactListError)? != '[' ||
-           s.chars().last().ok_or(ParseExactListError)? != ']'{
-            return Err(ParseExactListError);
-        }
-        // First, split the string into the sections between top level commas
-        let mut sections = Vec::<&str>::new();
-        let mut lastcomma = 0;
-        let mut depth = 0;
-        for (i, c) in s.chars().enumerate(){
-            if i == 0 {
-                continue;
-            } else if i == s.chars().count()-1 {
-                if depth != 0{
-                    return Err(ParseExactListError);
-                }
-                sections.push(&s[lastcomma+1..i]);
-                continue;
-            }
-
-            if c == '['{
-                depth += 1;
-            }else if c == ']'{
-                depth -= 1;
-                if depth < 0{
-                    return Err(ParseExactListError);
-                }
-            }else if c == ',' && depth == 0{
-                sections.push(&s[lastcomma+1..i]);
-                lastcomma = i;
-            }
-        }
-        // parse into numbers, lists and then a string if possible
-        let values = sections.iter().map(|ss| (*ss).trim().to_owned()).map(|ss|
-            if let Ok(n) = ss.parse() {
-                Value::Number(n)
-            } else if let Ok(l) = ss.parse::<ExactList>() {
-                Value::List(Rc::new(l))
-            } else if let Ok(l) = string::string_to_list(&ss) {
-                l
-            } else {
-                panic!("Unknown error parsing string \"{}\" in ExactList parser", ss);
-            }
-        );
-
-        return Ok(ExactList{
-            contents: values.collect()
-        })
-    }
-}
-
-impl LazyInductionList {
-    pub fn new(f: ParseTree, init: Value) -> LazyInductionList {
-        LazyInductionList {
-            function: f,
-            initial_value: init,
-            resolved: RefCell::new(Vec::new()),
-        }
-    }
-}
-
-impl ListLike for LazyInductionList {
-    fn index(&self, i: i64) -> Result<Value, RuntimeError> {
-        if i >= 0 {
-            let i = usize::try_from(i)
-                .map_err(|_| RuntimeError::OutOfBounds(format!("unknown error when indexing list (i = {})", i)))?;
-
-            let mut resolved = self.resolved.borrow_mut();
-            if resolved.len() == 0 {
-                resolved.push(self.initial_value.clone());
-            }
-            while i >= resolved.len() {
-                let prevresolved = resolved[resolved.len() - 1].clone();
-                resolved.push(evaluate::evaluate(&self.function, &prevresolved)?);
-            }
-            Ok(resolved[i].clone())
-        }else{
-            // negative indecies always return the first reached fixed point
-            // Err(RuntimeError::NegativeIndex(String::from("cannot negatively index infinite lists")))
-            loop {
-                let mut resolved = self.resolved.borrow_mut();
-                if resolved.len() == 0 {
-                    resolved.push(self.initial_value.clone());
-                }
-                let prevresolved = resolved[resolved.len() - 1].clone();
-                let nextresolved = evaluate::evaluate(&self.function, &prevresolved)?;
-                if prevresolved == nextresolved {
-                    return Ok(nextresolved);
-                }else{
-                    resolved.push(nextresolved);
-                }
-            }
-        }
-    }
-
-    fn length(&self) -> Result<i64, RuntimeError> {
-        return Err(RuntimeError::ResolvingInfiniteList(String::from("Cannot get length of infinite list")));
-    }
-
-    fn force_resolve(&self) -> Result<(), RuntimeError> {
-        Err(RuntimeError::ResolvingInfiniteList("Attempted to force_resolve an infinite list. (Does your final output include one?)".to_owned()))
-    }
-}
-
-impl LazyMapList {
-    pub fn new(f: ParseTree, s: Rc<dyn ListLike>) -> LazyMapList {
-        LazyMapList {
-            function: f,
-            source: s,
-            resolved: RefCell::new(Vec::new())
-        }
-    }
-}
-
-impl ListLike for LazyMapList {
-    fn index(&self, i: i64) -> Result<Value, RuntimeError> {
-        let trueindex;
-        let len = self.source.length()?;
-
-        if i >= len || i < -len{
-            return Err(RuntimeError::OutOfBounds(format!(
-                "Attempted to access index {} of list of length {}",
-                i,
-                len
-            )));
-        }else if i >= 0{
-            trueindex = i;
-        }else{
-            trueindex = len+i;
-        }
-
-        let trueindex = usize::try_from(trueindex)
-            .map_err(|_| RuntimeError::OutOfBounds(format!("unknown error when indexing list (i = {})", i)))?;
-
-        let mut resolved = self.resolved.borrow_mut();
-        while resolved.len() <= trueindex{
-            resolved.push(None);
-        }
-
-        Ok(
-            match &resolved[trueindex] {
-                None => {
-                    let ans = self.source
-                            .index(i)
-                            .and_then(|v| evaluate::evaluate(&self.function, &v))?;
-                    resolved[trueindex] = Some(ans.clone());
-                    ans
-                }
-                Some(ans) => ans.clone()
-            }
-        )
-    }
-
-    fn length(&self) -> Result<i64, RuntimeError> {
-        return self.source.length();
-    }
-
-    fn force_resolve(&self) -> Result<(), RuntimeError> {
-        let len = self.source.length()?;
-        self.source.force_resolve()?;
-
-        let mut resolved = self.resolved.borrow_mut();
-        while resolved.len() <= len as usize{
-            resolved.push(None);
-        }
-
-        for i in 0..len{
-            let trueindex = usize::try_from(i)
-                .map_err(|_| RuntimeError::OutOfBounds(format!("unknown error when indexing list (i = {})", i)))?;
-
-            resolved[trueindex] = Some(self.source
-                .index(i)
-                .and_then(|v| evaluate::evaluate(&self.function, &v))?);
-        }
-        Ok(())
-    }
-}
-
-impl LazyConcatList {
-    pub fn new(l1: Rc<dyn ListLike>, l2: Rc<dyn ListLike>) -> LazyConcatList {
-        let fl = l1.length()
-            .and_then(|n| i64::try_from(n).map_err(|_| RuntimeError::OutOfBounds(String::from("not sure what happened here"))))
-            .ok();
-        LazyConcatList {
-            first: l1,
-            second: l2,
-            firstlen: fl,
-        }
-    }
-}
-
-impl ListLike for LazyConcatList {
-    fn index(&self, i: i64) -> Result<Value, RuntimeError> {
-        match self.firstlen {
-            None => self.first.index(i),
-            Some(len) => {
-                if i < len {
-                    self.first.index(i)
-                } else {
-                    self.second.index(i - len)
-                }
-            }
-        }
-    }
-
-    fn length(&self) -> Result<i64, RuntimeError> {
-        return Ok(self.first.length()? + self.second.length()?);
-    }
-
-    fn force_resolve(&self) -> Result<(), RuntimeError> {
-        self.first.force_resolve()?;
-        self.second.force_resolve()
-    }
-}
-
 #[cfg(test)]
 mod tests {
+    use crate::parsetree::ParseTree;
     use super::*;
 
     #[test]
